+++ src/main/java/com/chinaunicom/rpc/entity/Task.java
@@ -2,23 +2,23 @@ package com.chinaunicom.rpc.entity;
 
 public class Task<T> {
 
-    private Long id;
+    private Integer id;
 
     private ServerThread serverThread;
 
     private T data;
 
-    public Task(Long id,T data,ServerThread serverThread){
+    public Task(Integer id,T data,ServerThread serverThread){
         this.id = id ;
         this.data = data;
         this.serverThread = serverThread;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
